feat: split package into cli and core
