+++ .eslintrc.json
@@ -1,1 +1,1 @@
-{ "root": true, "extends": ["@willbooster/eslint-config-ts"] }
+{ "root": true, "extends": [] }
