+++ .releaserc.json
@@ -4,6 +4,9 @@
     ["@semantic-release/commit-analyzer", { "preset": "conventionalcommits" }],
     "@semantic-release/release-notes-generator",
     "@semantic-release/npm",
-    "@semantic-release/github"
+    [
+      "@semantic-release/github",
+      { "successComment": false, "labels": ["r: semantic-release"], "releasedLabels": ["released :bookmark:"] }
+    ]
   ]
 }
