+++ package.json
@@ -1,83 +1,40 @@
 {
   "name": "firebase-scripts",
   "version": "0.0.0-semantically-released",
-  "description": "A collection of useful scripts for firebase.",
-  "keywords": [
-    "firebase"
-  ],
+  "private": true,
   "repository": "github:WillBooster/firebase-scripts",
   "license": "Apache-2.0",
   "author": "WillBooster Inc.",
-  "main": "dist/index.min.cjs",
-  "module": "dist/index.min.mjs",
-  "types": "dist/index.d.ts",
-  "bin": "./bin/main.js",
-  "files": [
-    "bin/",
-    "dist/"
+  "workspaces": [
+    "packages/*"
   ],
   "scripts": {
-    "build": "node esbuild.js",
+    "build": "yarn workspaces foreach --parallel --verbose run build",
     "cleanup": "yarn format && yarn lint-fix",
-    "format": "sort-package-json && yarn prettify",
+    "format": "sort-package-json && yarn prettify && yarn workspaces foreach --parallel --verbose run format",
     "postinstall": "husky install",
-    "lint": "eslint --color \"./{src,__tests__,scripts}/**/*.{cjs,js,jsx,mjs,ts,tsx}\"",
-    "lint-fix": "yarn lint --fix",
-    "prepack": "pinst --disable",
-    "postpack": "pinst --enable",
-    "prettify": "prettier --color --write \"**/{.*/,}*.{cjs,css,htm,html,js,json,json5,jsx,md,mjs,scss,ts,tsx,vue,yaml,yml}\" \"!**/test-fixtures/**\"",
-    "start": "node -r esbuild-register src/main.ts",
-    "stop-firebase": "kill-port --port 8080",
-    "test": "yarn test-wo-firebase && yarn test-w-firebase",
-    "test-w-firebase": "yarn stop-firebase && dotenv -- yarn test-w-firebase-core",
-    "test-w-firebase-core": "firebase emulators:exec --project $FIREBASE_PROJECT_ID \"yarn jest --silent --color --runInBand --forceExit __tests__/with-firebase\"",
-    "test-wo-firebase": "dotenv -- yarn jest --color __tests__/without-firebase",
-    "typecheck": "tsc --noEmit --Pretty"
+    "lint": "yarn workspaces foreach --parallel --verbose run lint",
+    "lint-fix": "yarn workspaces foreach --parallel --verbose run lint-fix",
+    "prettify": "prettier --color --write \"**/{.*/,}*.{cjs,css,htm,html,js,json,json5,jsx,md,mjs,scss,ts,tsx,vue,yaml,yml}\" \"!**/packages/**\" \"!**/test-fixtures/**\"",
+    "test": "yarn workspaces foreach --verbose run test",
+    "typecheck": "yarn workspaces foreach --parallel --verbose run typecheck"
   },
   "prettier": "@willbooster/prettier-config",
-  "resolutions": {
-    "npm/chalk": "^4.1.2"
-  },
-  "dependencies": {
-    "dayjs": "1.11.5",
-    "dotenv": "16.0.2",
-    "firebase-admin": "11.0.1",
-    "minimal-promise-pool": "1.1.5",
-    "yargs": "17.5.1"
-  },
   "devDependencies": {
-    "@firebase/rules-unit-testing": "2.0.4",
     "@types/eslint": "8.4.6",
     "@types/jest": "29.0.0",
     "@types/micromatch": "4.0.2",
     "@types/prettier": "2.7.0",
-    "@types/yargs": "17.0.12",
-    "@typescript-eslint/eslint-plugin": "5.36.1",
     "@typescript-eslint/parser": "5.36.1",
-    "@willbooster/eslint-config-ts": "9.2.1",
     "@willbooster/prettier-config": "9.1.0",
     "@willbooster/renovate-config": "9.4.0",
     "conventional-changelog-conventionalcommits": "5.0.0",
-    "dotenv-cli": "6.0.0",
-    "esbuild": "0.15.6",
-    "esbuild-plugin-d.ts": "1.1.0",
-    "esbuild-register": "3.3.3",
     "eslint": "8.23.0",
-    "eslint-config-prettier": "8.5.0",
-    "eslint-import-resolver-typescript": "3.5.0",
-    "eslint-plugin-import": "2.26.0",
-    "eslint-plugin-sort-class-members": "1.14.1",
-    "eslint-plugin-sort-destructure-keys": "1.4.0",
-    "firebase": "9.9.3",
-    "firebase-tools": "11.8.0",
     "husky": "8.0.1",
-    "jest": "29.0.1",
-    "kill-port": "2.0.1",
     "lint-staged": "13.0.3",
     "micromatch": "4.0.5",
     "pinst": "3.0.0",
     "prettier": "2.7.1",
-    "semantic-release": "19.0.5",
     "sort-package-json": "1.57.0",
     "typescript": "4.8.2"
   },
