+++ packages/cli/.eslintignore
