+++ packages/cli/.prettierignore
