+++ packages/cli/bin/main.js
