+++ packages/cli/package.json
@@ -1,0 +1,66 @@
+{
+  "name": "@firebase-scripts/cli",
+  "version": "0.0.0-semantically-released",
+  "description": "A cli of useful scripts for firebase.",
+  "keywords": [
+    "firebase"
+  ],
+  "repository": "github:WillBooster/firebase-scripts",
+  "license": "Apache-2.0",
+  "author": "WillBooster Inc.",
+  "main": "dist/index.min.cjs",
+  "module": "dist/index.min.mjs",
+  "types": "dist/index.d.ts",
+  "bin": "./bin/main.js",
+  "files": [
+    "bin/",
+    "dist/"
+  ],
+  "scripts": {
+    "build": "node esbuild.js",
+    "cleanup": "yarn format && yarn lint-fix",
+    "format": "sort-package-json && yarn prettify",
+    "lint": "eslint --color \"./{src,__tests__,scripts}/**/*.{cjs,js,jsx,mjs,ts,tsx}\"",
+    "lint-fix": "yarn lint --fix",
+    "prettify": "prettier --color --write \"**/{.*/,}*.{cjs,css,htm,html,js,json,json5,jsx,md,mjs,scss,ts,tsx,vue,yaml,yml}\" \"!**/test-fixtures/**\"",
+    "start": "dotenv -- node -r esbuild-register src/main.ts",
+    "typecheck": "tsc --noEmit --Pretty"
+  },
+  "prettier": "@willbooster/prettier-config",
+  "dependencies": {
+    "firebase-admin": "11.0.1",
+    "yargs": "17.5.1"
+  },
+  "devDependencies": {
+    "@firebase-scripts/core": "0.0.0-semantically-released",
+    "@firebase-scripts/shared": "0.0.0-semantically-released",
+    "@types/eslint": "8.4.6",
+    "@types/micromatch": "4.0.2",
+    "@types/prettier": "2.7.0",
+    "@types/yargs": "17.0.12",
+    "@typescript-eslint/eslint-plugin": "5.36.1",
+    "@typescript-eslint/parser": "5.36.1",
+    "@willbooster/eslint-config-ts": "9.2.1",
+    "@willbooster/prettier-config": "9.1.0",
+    "@willbooster/renovate-config": "9.4.0",
+    "conventional-changelog-conventionalcommits": "5.0.0",
+    "dotenv-cli": "6.0.0",
+    "esbuild": "0.15.6",
+    "esbuild-register": "3.3.3",
+    "eslint": "8.23.0",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-import-resolver-typescript": "3.5.0",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-sort-class-members": "1.14.1",
+    "eslint-plugin-sort-destructure-keys": "1.4.0",
+    "lint-staged": "13.0.3",
+    "micromatch": "4.0.5",
+    "prettier": "2.7.1",
+    "sort-package-json": "1.57.0",
+    "typescript": "4.8.2"
+  },
+  "packageManager": "yarn@4.0.0-rc.15",
+  "engines": {
+    "node": ">=14"
+  }
+}
