+++ packages/cli/src/cli.ts
