+++ packages/cli/src/copyCommand.ts
@@ -1,7 +1,6 @@
-import { app } from 'firebase-admin';
+import { copyDocument } from '@firebase-scripts/core/src/copy';
+import { initializeAdmin } from '@firebase-scripts/shared/src/firebaseAdmin';
 import type { CommandModule, InferredOptionTypes } from 'yargs';
-
-import { initializeAdmin } from './firebaseAdmin';
 
 const builder = {
   'source-service-account-path': {
@@ -42,21 +41,3 @@ export const copyCommand: CommandModule<unknown, InferredOptionTypes<typeof buil
     await copyDocument(srcAdminApp, source, destAdminApp, target);
   },
 };
-
-export async function copyDocument(
-  sourceAdminApp: app.App,
-  sourceDocumentPath: string,
-  targetAdminApp: app.App,
-  targetDocumentPath: string
-): Promise<void> {
-  console.info(`Reading the source document '${sourceDocumentPath}'...`);
-  const doc = (await sourceAdminApp.firestore().doc(sourceDocumentPath).get()).data();
-
-  if (!doc) {
-    console.error(`The source document '${sourceDocumentPath}' does not exist.`);
-    return;
-  }
-
-  console.info(`Writing the target document '${targetDocumentPath}'...'`);
-  await targetAdminApp.firestore().doc(targetDocumentPath).set(doc);
-}
