+++ packages/cli/src/main.ts
