+++ packages/core/.browserslistrc
@@ -1,0 +1,1 @@
+node 14
