+++ packages/core/.eslintignore
@@ -4,10 +4,13 @@
 # Generated by @willbooster/willboosterify
 
 3rd-party/
+@types/
+__generated__/
 android/
 ios/
 no-format/
 test-fixtures/
+*.config.js
 *.d.ts
 *.min.js
 .yarn/
@@ -503,3 +506,17 @@ dist
 .svelte-kit
 
 # End of https://www.toptal.com/developers/gitignore/api/node
+
+# Created by https://www.toptal.com/developers/gitignore/api/firebase
+# Edit at https://www.toptal.com/developers/gitignore?templates=firebase
+
+### Firebase ###
+# .idea
+**/node_modules/*
+**/.firebaserc
+
+### Firebase Patch ###
+.runtimeconfig.json
+.firebase/
+
+# End of https://www.toptal.com/developers/gitignore/api/firebase
