+++ packages/core/.gitignore
@@ -3,17 +3,6 @@
 
 # Generated by @willbooster/willboosterify
 
-3rd-party/
-android/
-ios/
-no-format/
-test-fixtures/
-*.d.ts
-*.min.js
-.yarn/
-.pnp.js
-
-
 .devcontainer/
 dist/
 temp/
@@ -309,6 +298,7 @@ flycheck_*.el
 # network security
 /network-security.data
 
+
 # End of https://www.toptal.com/developers/gitignore/api/emacs
 
 # Created by https://www.toptal.com/developers/gitignore/api/vim
@@ -503,3 +493,17 @@ dist
 .svelte-kit
 
 # End of https://www.toptal.com/developers/gitignore/api/node
+
+# Created by https://www.toptal.com/developers/gitignore/api/firebase
+# Edit at https://www.toptal.com/developers/gitignore?templates=firebase
+
+### Firebase ###
+# .idea
+**/node_modules/*
+**/.firebaserc
+
+### Firebase Patch ###
+.runtimeconfig.json
+.firebase/
+
+# End of https://www.toptal.com/developers/gitignore/api/firebase
