+++ packages/core/__tests__/without-firebase/convertObjectToTimestamp.test.ts
