+++ packages/core/firebase.json
