+++ packages/core/firebase/firestore.indexes.json
