+++ packages/core/firebase/firestore.rules
@@ -1,6 +1,6 @@
 rules_version = '2';
 service cloud.firestore {
   match /databases/{database}/documents {
-    allow read, write: if false;
+    allow read, write;
   }
 }
