+++ packages/core/jest.config.js
