+++ packages/core/package.json
@@ -1,0 +1,74 @@
+{
+  "name": "@firebase-scripts/core",
+  "version": "0.0.0-semantically-released",
+  "description": "A library of useful functions for firebase.",
+  "keywords": [
+    "firebase"
+  ],
+  "repository": "github:WillBooster/firebase-scripts",
+  "license": "Apache-2.0",
+  "author": "WillBooster Inc.",
+  "main": "dist/index.min.cjs",
+  "module": "dist/index.min.mjs",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist/"
+  ],
+  "scripts": {
+    "build": "node esbuild.js",
+    "cleanup": "yarn format && yarn lint-fix",
+    "format": "sort-package-json && yarn prettify",
+    "lint": "eslint --color \"./{src,__tests__,scripts}/**/*.{cjs,js,jsx,mjs,ts,tsx}\"",
+    "lint-fix": "yarn lint --fix",
+    "prettify": "prettier --color --write \"**/{.*/,}*.{cjs,css,htm,html,js,json,json5,jsx,md,mjs,scss,ts,tsx,vue,yaml,yml}\" \"!**/test-fixtures/**\"",
+    "stop-firebase": "kill-port --port 8080",
+    "test": "yarn test-wo-firebase && yarn test-w-firebase",
+    "test-w-firebase": "FIREBASE_PROJECT_ID=test; yarn stop-firebase && yarn test-w-firebase-core",
+    "test-w-firebase-core": "firebase emulators:exec --project $FIREBASE_PROJECT_ID \"yarn jest --silent --color --runInBand --forceExit __tests__/with-firebase\"",
+    "test-wo-firebase": "yarn jest --color __tests__/without-firebase",
+    "typecheck": "tsc --noEmit --Pretty"
+  },
+  "prettier": "@willbooster/prettier-config",
+  "dependencies": {
+    "minimal-promise-pool": "1.1.5"
+  },
+  "devDependencies": {
+    "@firebase-scripts/shared": "0.0.0-semantically-released",
+    "@firebase/rules-unit-testing": "2.0.4",
+    "@types/eslint": "8.4.6",
+    "@types/jest": "29.0.0",
+    "@types/micromatch": "4.0.2",
+    "@types/prettier": "2.7.0",
+    "@typescript-eslint/eslint-plugin": "5.36.1",
+    "@typescript-eslint/parser": "5.36.1",
+    "@willbooster/eslint-config-ts": "9.2.1",
+    "@willbooster/prettier-config": "9.1.0",
+    "@willbooster/renovate-config": "9.4.0",
+    "conventional-changelog-conventionalcommits": "5.0.0",
+    "esbuild": "0.15.6",
+    "esbuild-plugin-d.ts": "1.1.0",
+    "eslint": "8.23.0",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-import-resolver-typescript": "3.5.0",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-sort-class-members": "1.14.1",
+    "eslint-plugin-sort-destructure-keys": "1.4.0",
+    "firebase": "9.9.3",
+    "firebase-admin": "11.0.1",
+    "firebase-tools": "11.8.0",
+    "jest": "29.0.1",
+    "kill-port": "2.0.1",
+    "lint-staged": "13.0.3",
+    "micromatch": "4.0.5",
+    "prettier": "2.7.1",
+    "sort-package-json": "1.57.0",
+    "typescript": "4.8.2"
+  },
+  "peerDependencies": {
+    "firebase-admin": ">=11"
+  },
+  "packageManager": "yarn@4.0.0-rc.15",
+  "engines": {
+    "node": ">=14"
+  }
+}
