+++ packages/core/src/export.ts
@@ -2,49 +2,10 @@ import fsp from 'fs/promises';
 import path from 'path';
 
 import { app, firestore } from 'firebase-admin';
-import type { CommandModule, InferredOptionTypes } from 'yargs';
 
-import { initializeAdmin } from './firebaseAdmin';
 import { CompressionFormat, compressJsonText, getExtensionFromFormat, getFormatFromExtension } from './jsonCompressor';
 
-export const DEFAULT_BATCH_SIZE = 10000;
-
-const builder = {
-  directory: {
-    type: 'string',
-    description: 'A directory path where serialized files are generated',
-    alias: 'd',
-  },
-  format: {
-    type: 'string',
-    description: 'A compression format (gzip or brotil)',
-    alias: 'f',
-  },
-  batchSize: {
-    type: 'number',
-    description: 'A batch size to fetch document from collection',
-    default: DEFAULT_BATCH_SIZE,
-    alias: 'b',
-  },
-} as const;
-
-export const exportCommand: CommandModule<unknown, InferredOptionTypes<typeof builder>> = {
-  command: 'export',
-  describe: 'Export and serialize specified collections',
-  builder,
-  async handler(argv) {
-    const adminApp = initializeAdmin();
-    await exportCollections(
-      adminApp,
-      argv._.map((arg) => arg.toString()),
-      argv.directory ?? path.resolve(),
-      {
-        format: argv.format as CompressionFormat,
-        batchSize: argv.batchSize,
-      }
-    );
-  },
-};
+export const DEFAULT_BATCH_SIZE = 1000;
 
 export interface ExportOptions {
   batchSize?: number;
