+++ packages/core/src/jsonCompressor.ts
