+++ packages/shared/.eslintignore
