+++ packages/shared/src/firebaseAdmin.ts
