+++ packages/shared/tsconfig.json
@@ -4,23 +4,14 @@
     "alwaysStrict": true,
     "esModuleInterop": true,
     "importHelpers": false,
-    "module": "es2015",
+    "module": "esnext",
     "moduleResolution": "node",
     "outDir": "dist",
     "resolveJsonModule": true,
     "skipLibCheck": true,
     "sourceMap": true,
     "strict": true,
-    "target": "es2019"
+    "target": "esnext"
   },
-  "include": [
-    "src/**/*",
-    "__tests__/**/*",
-    "scripts/**/*",
-    "packages/*/src/**/*",
-    "packages/*/__tests__/**/*",
-    "packages/*/scripts/**/*",
-    "./node_modules/@types",
-    "./@types"
-  ]
+  "include": ["src/**/*", "__tests__/**/*", "scripts/**/*", "../../node_modules/@types", "../../@types", "./@types"]
 }
